Initialize Maker in POST /swaps
