+++ maker/.env.sample
@@ -1,6 +1,9 @@
-BITCOIN_KEY=tprv8ZgxMBicQKsPdcw4s6QKZoCPLyQV46dbg5MDPg7AS6E4ucy9hedENyJZWXoHEFSn3sNpsNtKR9SL8G5ERXc6rpxRjZkwMNfigkrDwLy35ZX
-BITCOIN_NODE_URI=127.0.0.1:63758
-ERC20_CONTRACT_ADDRESS=0x62d69f6867a0a084c6d313943dc22023bc263691
-ETHEREUM_KEY=1e8229295a37550ebc25228cafee593eb61d047bc4e1225cd8ae0195f1b70fd6
-ETHEREUM_NODE_URL=http://127.0.0.1:63743
-CND_NODE_URL=http://127.0.0.1:63766
+BITCOIN_HD_KEY_0=
+BITCOIN_HD_KEY_1=
+BITCOIN_P2P_URI=
+ERC20_CONTRACT_ADDRESS=
+ETHEREUM_KEY_0=
+ETHEREUM_KEY_1=
+ETHEREUM_NODE_HTTP_URL=
+HTTP_URL_CND_0=
+HTTP_URL_CND_1=
