+++ maker/services/swaps/comit.js
@@ -1,0 +1,53 @@
+const dotenv = require("dotenv");
+const {
+  Actor,
+  BigNumber,
+  createActor,
+  EthereumWallet,
+  InMemoryBitcoinWallet,
+  SwapRequest,
+} = require("comit-sdk");
+const fs = require("fs");
+const os = require("os");
+const path = require("path");
+
+function loadEnvironment() {
+  // TODO: change this to use .env within the `maker` directory
+  const envFilePath = path.join(os.homedir(), ".create-comit-app", "env");
+
+  if (!fs.existsSync(envFilePath)) {
+      console.log(
+          "Could not find file %s. Did you run `yarn start-env`?",
+          envFilePath
+      );
+      process.exit(1);
+  }
+
+  dotenv.config({ path: envFilePath });
+}
+
+async function initializeNodes(index, name) {
+  const bitcoinWallet = await InMemoryBitcoinWallet.newInstance(
+    "regtest",
+    process.env.BITCOIN_P2P_URI,
+    process.env[`BITCOIN_HD_KEY_${index}`]
+  );
+  await new Promise(r => setTimeout(r, 1000)); // TODO: why is this here
+
+  const ethereumWallet = new EthereumWallet(
+      process.env.ETHEREUM_NODE_HTTP_URL,
+      process.env[`ETHEREUM_KEY_${index}`]
+  );
+
+  return createActor(
+      bitcoinWallet,
+      ethereumWallet,
+      process.env[`HTTP_URL_CND_${index}`],
+      name
+  );
+}
+
+module.exports = {
+  loadEnvironment,
+  initializeNodes
+}
